Make hello test automatable
Took the very simple approach here of always running a single task
on a single node to avoid the need for sorting, adapting to a number
of cores, etc.  I figure these patterns are common enough that if
this keeps compiling and running, it'll keep working as intended
or lots of other things will break.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>